src/home-lights/requirements.txt
requests
python-kasa
asyncio-mqtt
